gravityfix keeps g_gravity cfg intead of forcing
// G_SpawnString( "gravity", "800", &s );	// MIKEAUS original code disabled, so everytime gameserver starts or restarts map it keeps g_gravity settings fixed in config cfg instead of forcing to 800
// trap_Cvar_Set( "g_gravity", s );		// MIKEAUS original code disabled, as above

gravity was always reset and sets to 800. game server is written the way that: default game gravity is set to 800, g_gravity is value set by each map or rcon to be adjusted. that means that acording to main code, each map has to setup its own gravity by g_gravity setting. not the server gravity (which is default 800), and this is the reason why it is restarting to 800 every single time.
now, everytime gameserver starts or restarts map it keeps g_gravity settings fixed in config cfg instead of forcing to 800 as default set